+++ test/cases/fetch_test_sqlserver.rb
@@ -42,6 +42,25 @@ class FetchTestSqlserver < ActiveRecord::TestCase
     end
   end
 
+  describe "FROM subquery" do
+    let(:from_sql) { "(SELECT [books].* FROM [books]) [books]" }
+
+    it "SQL generated correctly for FROM subquery if order provided" do
+      query = Book.from(from_sql).order(:id).limit(5)
+
+      assert_equal query.to_sql, "SELECT [books].* FROM (SELECT [books].* FROM [books]) [books] ORDER BY [books].[id] ASC OFFSET 0 ROWS FETCH NEXT 5 ROWS ONLY"
+      assert_equal query.to_a.count, 5
+    end
+
+    it "exception thrown if FROM subquery is provided without an order" do
+      query = Book.from(from_sql).limit(5)
+
+      assert_raise(ActiveRecord::StatementInvalid) do
+        query.to_sql
+      end
+    end
+  end
+
   protected
 
   def create_10_books